Refactory e teste integrado
